Initial project finishing commit - Refactored accounts to use BigDecimal in the database and pull the correct data - Added both reports (generateStatement and generateAccountStatement) - Added Transaction object - Added transactions handling for withdrawal/deposit functions and simulating a transaction - Updated ErrorUtil to ErrorHandlingUtil - Added gson for file writing into build.gradle.kts - Added LocalDate to capture date for database entry + report creation

+++ banking-ledger-system/src/main/java/edu/mcc/codeschool/objects/Account.java
@@ -1,9 +1,11 @@
 package edu.mcc.codeschool.objects;
 
+import java.math.BigDecimal;
+
 public class Account {
     private Integer accountID;
     private Integer CustomerID;
-    private Double balance;
+    private BigDecimal balance;
     private String name;
     private Long accountNumber;
     private String customerName;
@@ -26,11 +28,11 @@ public class Account {
         return this;
     }
 
-    public Double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public Account setBalance(Double balance) {
+    public Account setBalance(BigDecimal balance) {
         this.balance = balance;
         return this;
     }
